9(1)-RAG/Rag-session/requirements.txt
streamlit>=1.30.0
datasets>=2.16.0
elasticsearch>=8.12.0,<9
pinecone>=3.0.0
openai>=1.10.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.66.0
